# Modify "get_user_config_dir()" function