--- a/ultralytics/utils/__init__.py
+++ b/ultralytics/utils/__init__.py
@@ -774,7 +774,7 @@ def get_ubuntu_version():
             return None
 
 
-def get_user_config_dir(sub_dir="Ultralytics"):
+def get_user_config_dir(sub_dir="yolov12"):
     """
     Return the appropriate config directory based on the environment operating system.