Thank you for your work.

Resolve the issue where the output of the yolov12 project is being generated in the 'Ultralytics' folder when both 'Ultralytics' and 'yolov12' folders exist on the PC.

- Before change
```
$ yolo detect predict model=./yolov12x.pt
FlashAttention is not available on this device. Using scaled_dot_product_attention instead.
/home/test/anaconda3/envs/yolov12/lib/python3.11/site-packages/timm/models/layers/__init__.py:48: FutureWarning: Importing from timm.models.layers is deprecated, please import via timm.layers
  warnings.warn(f"Importing from {__name__} is deprecated, please import via timm.layers", FutureWarning)
WARNING ⚠️ 'source' argument is missing. Using default 'source=/home/test/Desktop/YOLO/yolov12/ultralytics/assets'.
Ultralytics 8.3.63 🚀 Python-3.11.11 torch-2.2.2+cu121 CUDA:0 (Quadro RTX 4000, 7952MiB)
YOLOv12x summary (fused): 626 layers, 59,135,744 parameters, 0 gradients, 199.0 GFLOPs

image 1/2 /home/test/Desktop/YOLO/yolov12/ultralytics/assets/bus.jpg: 640x480 4 persons, 1 bus, 98.8ms
image 2/2 /home/test/Desktop/YOLO/yolov12/ultralytics/assets/zidane.jpg: 384x640 2 persons, 2 ties, 102.1ms
Speed: 2.2ms preprocess, 100.5ms inference, 276.2ms postprocess per image at shape (1, 3, 384, 640)
Results saved to /home/test/Desktop/ultralytics/runs/detect/predict
💡 Learn more at https://docs.ultralytics.com/modes/predict
```

- After change
```
$ yolo detect predict model=./yolov12x.pt
FlashAttention is not available on this device. Using scaled_dot_product_attention instead.
/home/test/anaconda3/envs/yolov12/lib/python3.11/site-packages/timm/models/layers/__init__.py:48: FutureWarning: Importing from timm.models.layers is deprecated, please import via timm.layers
  warnings.warn(f"Importing from {__name__} is deprecated, please import via timm.layers", FutureWarning)
WARNING ⚠️ 'source' argument is missing. Using default 'source=/home/test/Desktop/YOLO/yolov12/ultralytics/assets'.
Ultralytics 8.3.63 🚀 Python-3.11.11 torch-2.2.2+cu121 CUDA:0 (Quadro RTX 4000, 7952MiB)
YOLOv12x summary (fused): 626 layers, 59,135,744 parameters, 0 gradients, 199.0 GFLOPs

image 1/2 /home/test/Desktop/YOLO/yolov12/ultralytics/assets/bus.jpg: 640x480 4 persons, 1 bus, 100.0ms
image 2/2 /home/test/Desktop/YOLO/yolov12/ultralytics/assets/zidane.jpg: 384x640 2 persons, 2 ties, 103.7ms
Speed: 2.1ms preprocess, 101.9ms inference, 242.0ms postprocess per image at shape (1, 3, 384, 640)
Results saved to /home/test/Desktop/YOLO/yolov12/runs/detect/predict
💡 Learn more at https://docs.ultralytics.com/modes/predict
```